Moving core functions into the 'haro' class
Moving core functions into the 'haro' class
