+++ CHANGELOG.md
@@ -25,7 +25,7 @@ Generated by [`auto-changelog`](https://github.com/CookPete/auto-changelog).
 
 - Updating build [`#73`](https://github.com/avoidwork/haro/pull/73)
 - Removing issues [`21ccad7`](https://github.com/avoidwork/haro/commit/21ccad7676a7f1e2f2490fc6bc7b9d48f657821a)
-- Updating build tools, adding constants.js, updating utils.js, setting 'this.key' if not on the record [`646e655`](https://github.com/avoidwork/haro/commit/646e6553f38d41aa24118b560ab207e4d737cd47)
+- Updating build tools, adding constants.js, updating uuid.js, setting 'this.key' if not on the record [`646e655`](https://github.com/avoidwork/haro/commit/646e6553f38d41aa24118b560ab207e4d737cd47)
 - Building outputs [`9980d75`](https://github.com/avoidwork/haro/commit/9980d7590337b83f03d2759a995768b737ee9d71)
 
 #### [14.0.1](https://github.com/avoidwork/haro/compare/14.0.0...14.0.1)
